+++ src/tests/cmd_hook_entry.rs
